create formatToLocalTimeZone util file and use it for home page and events
